updates for python3

+++ rna_tools/sym_link.py
@@ -19,7 +19,7 @@ CWD = getcwd()
 chdir( dirname( argv[0] ) )
 
 for f in glob.glob('./bin/*'):
-    os.remove(f)
+    if not os.path.isdir(f): os.remove(f)
 
 for dirpath, dirnames, filenames in os.walk('./'):
     if len( basename(dirpath) ) == 0 or basename(dirpath)[0] == '.' or basename(dirpath) == 'bin':
@@ -32,7 +32,7 @@ for dirpath, dirnames, filenames in os.walk('./'):
             file2 = './bin/' + f
             try:
                 os.symlink(file1, file2)
-            except OSError, e:
+            except (OSError, e):
                 if e.errno == errno.EEXIST:
                     os.remove(file2)
                     os.symlink(file1, file2)
